New algorithms ABST_UWFS, DIFFUSIVE_WAVE_VARY; Bug fixed for CMM [BENCHMARKED]
Added ABST_UWFS algorithm (Abstraction.cpp/DepressionProcesses.h)
-new surface params UWFS_B and UWFS_BETAMIN (LandUseClass.cpp; Properties.h)
-parse support (ParseInput.cpp)

Added DIFFUSIVE_WAVE_VARY routing algorithm (ModelParamCheck.cpp;RavenInclude.h, sUbBasin.cpp/.h)
-parse support (ParseInput.cpp)
-new routine CSubBasin::UpdateRoutingHydro()
-new CSubBasin member _c_hist

Improved mass balance support for diversions in CModel::IncrementCumOutflow (Model.cpp; Model.h; RavenMain.cpp)

Moved snow cover depletion curve use to CHydroUnit::GetSnowCover() (HydroUnits.cpp/SnowParams.cpp)
-intialized Options.snow_depletion in ParseInput.cpp

Unfinished beginnings:
-Upgrading SurfaceEnergyExchange.cpp routine - still needs work
---added new vegetation parameter VEG_CONV_COEFF (Properties.h; VegetationClass.cpp)
-Started filling out ParseNetCDFStateFile() (ParseFEWSRunInfo.cpp)

QA/QC
-added isnan check to ForcingGrid.cpp

Bug fix
-fixed destructor memory error in ForcingGrid for simulations not ending at 00:00 found by Mabrouk@CMM - resized _aVal matrix in ForcingGrid.cpp
-fixed Output Dump issue if exceeding num output times found by Mabrouk@CMM (StandardOutput.cpp)

Unit Testing
-added TestManningsInfluence() to ChannelXSect.cpp

Cleanup
-renamed type member variable to _type in DepressionProcesses.h/Abstraction.cpp/DepressionProcesses.cpp
-minor recast to strlen in CommonFunctions.cpp to avoid compiler warnings
-added deletion of small character array unit_t in ForcingGrid.cpp
-removed all use of unused Options.distrib_lat_inflow option (MassRouting.cpp/RavenInclude.h/SubBasin.cpp)
-deleted unused soil parameter compressibility
-removed unused PDMROF_b reference in SOILEVAP_HYPR (SoilEvaporation.cpp)
-deleted unused files SoilWaterMovers2.cpp and PropertyClassABC.h

git-svn-id: https://chyms.nrc.gc.ca/svn/raven/trunk@302 9a3899ca-7adb-4d27-9b38-7f4ebb3123b0

+++ src/CommonFunctions.cpp
@@ -486,11 +486,11 @@ time_struct TimeStructFromNetCDFString(const string unit_t_str,const string time
   if(!strstr(tmp.substr(start+7 ,1).c_str(),"-")) { tmp.insert(start+5,"0"); } // second dash in date - fixes YYYY-M-dd
 
   bool date_only=false;
-  if (strlen(tmp.c_str())<=(start+10+2)){date_only=true;} //00:00:00 +0000 not included in string
+  if ((int)(strlen(tmp.c_str()))<=(start+10+2)){date_only=true;} //00:00:00 +0000 not included in string
 
   if(!date_only) {
     if(!strstr(tmp.substr(start+10,1).c_str()," ") && !strstr(tmp.substr(start+10,1).c_str(),"T")) { tmp.insert(start+8,"0"); } // second dash in date - fixes YYYY-MM-d
-    if(strstr(tmp.substr(start+10,1).c_str(),"T")) {tmp.replace(start+10,1," "); } // replacing 'T' with ' ' in case date looks like YYYY-MM-DDTHH:MM:SS
+    if( strstr(tmp.substr(start+10,1).c_str(),"T")) {tmp.replace(start+10,1," "); } // replacing 'T' with ' ' in case date looks like YYYY-MM-DDTHH:MM:SS
     if(!strstr(tmp.substr(start+13,1).c_str(),":")) { tmp.insert(start+11,"0"); } // first colon in time  - fixes 1:00:00
     if(!strstr(tmp.substr(start+16,1).c_str(),":")) { tmp.insert(start+14,"0"); } // second colon in time - fixes 11:0:00 (?)
   }
